Separator structure
